module-5/hw-9/fastapi/requirements.txt
fastapi>=0.95.0
python-multipart
uvicorn
torch
torchvision
pytorch-lightning
Pillow
jinja2
pytest